removed redundant node_modules

+++ news_app/api/services/news_service.py
@@ -32,18 +32,21 @@ def get_news_by_uuid(uuid: str) -> t.Optional[News]:
     return news
 
 
-def update_news(data: dict) -> int:
+def update_news(data: dict, user: User) -> int:
     photo = data.pop('photo', None)
     tags = data.pop('tags', [])
     rubrics = data.pop('rubrics', [])
     uuid = data.pop('uuid', None)
-    result = News.objects.filter(uuid=uuid).update(**data)
-    print(result)
+
+    result = News.objects.filter(uuid=uuid, author_id=user.uuid).update(**data)
     if not result:
         return result
 
+    news_entity = News.objects.filter(uuid=uuid, author_id=user.uuid).first()
+    if not news_entity:
+        return 0
+
     # updating relationships
-    news_entity = News.objects.filter(uuid=uuid).first()
     news_entity.photo = photo
     news_entity.tags.set(tags)
     news_entity.rubrics.set(rubrics)
@@ -52,10 +55,9 @@ def update_news(data: dict) -> int:
     return result
 
 
-def delete_news(uuid: str) -> int:
-    print(uuid)
+def delete_news(uuid: str, user: User) -> int:
     if not uuid:
         return 0
-    res = News.objects.filter(uuid=uuid).delete()
+    res = News.objects.filter(uuid=uuid, author_id=user.uuid).delete()
     deleted_objects_count = res[0]
     return deleted_objects_count
